+++ tests/test_gridkit/test_hex_grid.py
@@ -21,53 +21,29 @@ def test_centroid(shape, indices, expected_centroids):
     numpy.testing.assert_allclose(centroids, expected_centroids)
 
 
-@pytest.mark.parametrize("shape, bounds, expected_ids", [
-    ["pointy", (0,-2,2,2), [[0,0],[-1,-1]]],
-    ["pointy", (0,-2,3,2), [[0,0],[-1,-1],[0,-1]]],
-    ["pointy", (0,0,1,5), [[-1,1],[0,0]]],
-    ["pointy", (0,0,3,5), [[-1,1],[0,1],[0,0]]],
-    ["pointy", (-2,-2,2,2), [[-1,0],[0,0],[-1,-1]]],
-    ["pointy", (-2,-2,3,2), [[-1,0],[0,0],[-1,-1],[0,-1]]],
-    ["pointy", (-2,0,2,5), [[-1,1],[-1,0],[0,0]]],
-    ["pointy", (-2,0,3,5), [[-1,1],[0,1],[-1,0],[0,0]]],
-    ["flat", (-2,-2,2,2), [[-1,-1],[0,-1],[0,0]]],
-    ["flat", (-2,-2,2,3), [[-1,-1],[-1,0],[0,-1],[0,0]]],
-    ["flat", (0,-2,5,2), [[0,-1],[0,0],[1,-1]]],
-    ["flat", (0,-2,5,3), [[0,-1],[0,0],[1,-1],[1,0]]],
-    ["flat", (-2,0,2,2), [[-1,-1],[0,0]]],
-    ["flat", (-2,0,2,3), [[-1,-1],[-1,0],[0,0]]],
-    ["flat", (-2,2,2,3), [[-1,0],[0,0]]],
-    ["flat", (-2,2,2,5), [[-1,0],[0,0],[0,1]]],
-    ["flat", (0,0,5,1), [[0,0],[1,-1]]],
-    ["flat", (0,0,5,3), [[0,0],[1,-1],[1,0]]],
+@pytest.mark.parametrize("shape, bounds, expected_ids, expected_shape", [
+    ["pointy", (0,-2,2,2), [[-1,-1], [0,0]], (2,1)],
+    ["pointy", (0,-2,3,2), [[-1,-1],[0,0]], (2,1)],
+    ["pointy", (0,0,3,5), [[0,0],[-1,1]], (2,1)],
+    ["pointy", (-2,-2,2,2), [[-2,-1],[-1,-1],[-1,0],[0,0]], (2,2)],
+    ["pointy", (-2,-2,3,2), [[-2,-1],[-1,-1],[-1,0],[0,0]], (2,2)],
+    ["pointy", (-2,0,2,5), [[-1,0],[0,0],[-2,1],[-1,1]], (2,2)],
+    ["pointy", (-2,0,3,5), [[-1,0],[0,0],[-2,1],[-1,1]], (2,2)],
+    ["flat", (-2,-2,2,2), [[-1,-2],[-1,-1],[0,-1],[0,0]], (2,2)],
+    ["flat", (-2,-2,2,3), [[-1,-2],[-1,-1],[0,-1],[0,0]], (2,2)],
+    ["flat", (0,-2,5,2), [[1,-2],[0,-1],[1,-1],[0,0]], (2,2)],
+    ["flat", (0,-2,5,3), [[1,-2],[0,-1],[1,-1],[0,0]], (2,2)],
+    ["flat", (-2,0,2,2), [[-1,-1],[0,0]], (2,1)],
+    ["flat", (-2,0,2,3), [[-1,-1],[0,0]], (2,1)],
+    ["flat", (-2,2,2,5), [[-1,0],[0,1]], (2,1)],
+    ["flat", (0,0,5,3), [[1,-1],[0,0]], (2,1)],
 ])
-def test_cells_in_bounds(shape, bounds, expected_ids):
+def test_cells_in_bounds(shape, bounds, expected_ids, expected_shape):
     grid = HexGrid(size=3, shape=shape)
     aligned_bounds = grid.align_bounds(bounds, mode="nearest")
-    ids, _ = grid.cells_in_bounds(aligned_bounds) #TODO: remove returned None shape
-
-    import matplotlib.pyplot as plt
-    def bounds_to_poly(bounds): 
-        return (
-            [[bounds[0], bounds[0], bounds[2], bounds[2], bounds[0]],
-            [bounds[1], bounds[3], bounds[3], bounds[1], bounds[1]]]
-        )
-    fig, ax = plt.subplots()
-    ax.fill(
-            *bounds_to_poly(aligned_bounds),
-            alpha=0.5,
-            color="green"
-        )
-    for poly, id in zip(grid.to_shapely(ids), ids):
-        x, y = poly.exterior.xy
-        ax.fill(
-            x,y,
-            alpha=0.5,
-            color="red"
-        )
-        ax.text(*grid.centroid(id), id)
-    fig.savefig(f"test_{bounds}.png")
+    ids, shape = grid.cells_in_bounds(aligned_bounds)
     numpy.testing.assert_allclose(ids, expected_ids)
+    numpy.testing.assert_allclose(shape, expected_shape)
 
 
 @pytest.mark.parametrize("shape, point, expected_nearby_cells", [