+++ docs/source/release_notes.rst
@@ -3,3 +3,34 @@
 Release notes
 ================
 
+Version 0.2.0 (July 10, 2023)
+-----------------------------
+
+Features
+ - Add hex_grid.HexGrid class
+ - Add hex_grid.BoundedHexGrid class
+ - `to_shapely()` on bounded grids returns the shapes in the bounds when no index is supplied
+ - add action for pytest and doctest on push
+ - turn bounded_grid.indices into a property
+
+Fixes
+ - set proper version when documentation is build
+
+Documentation
+ - build documentation when tagged instead of merged in main
+ - add example "Hexagon grids"
+ - add example "Cell selection using other grids"
+ - add example "Resampling"
+ - use hexagons instead of squares in example "Interpolate from points"
+
+
+Version 0.1.1 (March 17, 2023)
+------------------------------
+
+Fixes
+ - Fix `__version__`` missing an ending quotation mark
+
+
+Version 0.1.0 (March 17, 2023)
+------------------------------
+ - release first version to PyPi